segment_INR_data

+++ mmseg/models/decode_heads/decode_head.py
@@ -241,7 +241,7 @@ class BaseDecodeHead(BaseModule, metaclass=ABCMeta):
             seg_weight = self.sampler.sample(seg_logit, seg_label)
         else:
             seg_weight = None
-        seg_label = seg_label.squeeze(1)
+        seg_label = seg_label.squeeze(1).long()
 
         if not isinstance(self.loss_decode, nn.ModuleList):
             losses_decode = [self.loss_decode]
